+++ geokey/socialinteractions/models.py
@@ -10,9 +10,9 @@ from allauth.socialaccount.models import SocialAccount, SocialToken, SocialApp
 import tweepy
 import facebook
 
-from geokey.contributions.models import Observation, Comment
+from geokey.subsets.models import Subset
+# from geokey.contributions.models import Observation, Comment
 from .base import STATUS, FREQUENCY
-from .utils import check_provider
 
 
 class SocialInteraction(models.Model):
@@ -65,7 +65,7 @@ class SocialInteractionPull(models.Model):
     updated_at = models.DateTimeField(null=True, auto_now_add=False)
 
 
-@receiver(post_save, sender=Observation)
+@receiver(post_save, sender=Subset)
 def post_social_media(sender, instance, created, **kwargs):
     """Post/tweet to social media when a new Observation is added.
 
@@ -116,3 +116,41 @@ def post_social_media(sender, instance, created, **kwargs):
                     commentto=instance,
                     creator=socialaccount.user
                 )
+
+
+def check_provider(provider, access_token, text_to_post, app):
+    """Check the provider and authenticate.
+
+    Parameters:
+    ------------
+    provider :  str
+        provider of the social account
+    access_token: str - SocialToken Object
+        access token for the social account and user
+    text_to_post: str
+        text which will be posted to social media
+    app: socialAccount app object
+
+    returns
+    --------
+    tweet_id : str
+        tweet identifier
+    screen_aname: str
+        screen name user by twitter user
+
+    """
+    if provider == 'facebook':
+        graph = facebook.GraphAPI(access_token)
+        graph.put_wall_post(message=text_to_post)
+    if provider == 'twitter':
+        consumer_key = app.client_id
+        consumer_secret = app.secret
+        auth = tweepy.OAuthHandler(consumer_key, consumer_secret)
+        access_token_all = access_token
+        access_token = access_token_all.token
+        access_token_secret = access_token_all.token_secret
+        auth.set_access_token(access_token, access_token_secret)
+        api = tweepy.API(auth)
+        tweet_back = api.update_status(text_to_post)
+
+    return tweet_back.id, tweet_back.author.screen_name
